refactor(Menu): 주문 메뉴의 문자열 변환 기능을 리팩토링함
변경 내용:
- newline을 상수화하고 Menu에 이를 반영
- 맨 끝의 newline은 OutputView에서 관리하도록 추가
